refactor(dashboard): update ClassroomModal component and move NewStudent component to same folder
