+++ packages/frontend/src/components/dashboard/modals/ClassroomModal/index.ts
@@ -1,0 +1,3 @@
+import ClassroomModal from './ClassroomModal';
+
+export default ClassroomModal;
