--- packages/frontend/src/components/dashboard/teacher/NewStudent/index.ts
@@ -1,3 +1,0 @@
-import NewStudent from './NewStudent';
-
-export default NewStudent;
